Organize file folders according to conventions
