+++ Content.Shared/Whistle/WhistleSystem.cs
@@ -1,4 +1,3 @@
-using Content.Shared.Whistle.Components;
 using Content.Shared.Coordinates;
 using Content.Shared.Interaction.Events;
 using Content.Shared.Movement.Components;
